requirements.txt
asgiref==3.8.1
Django==5.1.6
django-appconf==1.1.0
djangorestframework==3.15.2
drf-yasg==1.21.9
inflection==0.5.1
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pytest==8.3.5
pytest-django==4.10.0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.1.1
